jooq-generated code was moved to a separate module
